[api change] introducing `hoodie.id()`. `Replaces hoodie.account.ownerHash`.
**important:** This two breaking changes:

1. The now called hoodieId has up till now been locally stored in `$config/hoodie`
   as `_account.ownerHash`, but now as `_hoodieId`.
2. All /_users docs have the property `ownerHash`, it has now be renamed to `hoodieId`

That means that in order to update hoodie.js on current applications, both the CouchDB
and localStorage should be migrated. We do not have a feature build in for that at
this point, so it's in everyone's responsibility to do so. But we should consider to
add such a feature for the future.

* * *

This commit was sponsored by Digital Ocean.
http://go.hood.ie/hoodie-digital-ocean