[hooks] make sure we call all hooks for all registered plugins
The previous code would exit the looping over plugins early
if the current hook wasn’t found in the current plugin.

This also fixes the reporting of whether a hook as been run
to the calling code.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:
http://go.hood.ie/thehoodiefirm

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:
http://go.hood.ie/thehoodiefirm