revert: hoodie.account.anonymousSignUp depends on promise.progress
We have to fix it anyway. I will rather have it fail until we got it

* * *

This commit was sponsored by Toptal, a network of senior developers
who work with leading companies and teams across the globe. If you’re
looking to take your freelance work up a notch, check them out.

http://goo.gl/MtjLLA