feat(http): Allow clients to request a gzip response with ?gzip=force or cookie
Some clients, while understanding gzip resources and sending the
correct Accept-Encoding HTTP header are behind proxy servers that
strip the header, resulting the server to send a non-gzip response.

Steve Souders, Tony Gentilcore and Andy Martone outline a way to
resolve this in:

http://www.stevesouders.com/blog/2010/07/12/velocity-forcing-gzip-compression/

This implements the server-side component of this scheme. Any request
to /_api that has a query parameter `?gzip=force` or a cookie value
`hoodie_force_gzip=true` will receive a gzipped response, even if no
Accept-Encoding header is present or if it is mangled.

Thanks to Philip @latentflip Roberts for brainstorming an approach and
for providing a proof of concept that I could work off of.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:

http://go.hood.ie/thehoodiefirm