fix(test): fix hanging tests, speed up tests
We now only start one hoodie-server for all integration tests
(except for the `require` one). This speeds up the integration
test runtime. It also helps to not use too many sockets that are
used in HTTP requests between hoodie-server and CouchDB during
startup and setup.

In addition, test cases now use the `pool: false` parameter with
`http.request`, to avoid even more keepAlive issues.

This feels like treating a sympton and not the cause, but given
that this has a bunch of benefits, I’m inclined to include this.

The actual issue will likely be fixed in a future hoodie-server
refactor.