refactor: unravel hapi server configuration
This touches and changes a lot of things, so that the hapi server creation is now at the heart and center of this module (where it belongs).
Because the tests aren't really tests – they just check if certain internal functions exist – this has to touch and delete tests as well.
While that totally looks like it's breaking everything the surfacing API remains exactly the same.

The function signature of `require('hoodie-server')` changed, so that it doesn't auto start the hoodie-server, but it just passes it to the callback.
This way the CLI actually takes care of starting it, and some logging could already be separated out of the logic.
This will eventually make testing easier.

This is an API that was never documented or public and therefor this is not a breaking change.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:

http://go.hood.ie/thehoodiefirm

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:

http://go.hood.ie/thehoodiefirm