refactor(config_store): replace with plain simple sync get/set
The config_store module was essentially trying to take a sledgehammer to
crack a nut. We're saving to the `config.json` in only two ocassions, and
reading from it at few times. This is nothing that would require any file
locking and other over the top performance/failability checks. Let's care about
that again once it becomes a problem.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:

http://go.hood.ie/thehoodiefirm