feat: when no COUCH_URL is provided spawn pouchdb-server instead of couchdb
BREAKING CHANGE: The functionality to automatically start a CouchDB got
removed from Hoodie. If you want to run Hoodie with a CouchDB instance you
have to start it yourself and provide Hoodie the URL where it's reachable.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:

http://go.hood.ie/thehoodiefirm