fix: use 8080 as default hoodie port
Before this change hoodie generated and persisted a unique id
and derived a port from that using the `ports` package.
It's now simply using 8080 as the default port and using whatever
is provided with the port option. With the `.hoodierc` users may
persist their choice and have the same effect, but without the
overhead in `hoodie-server`. This change made the entire
`config-store.js` obsolete.