chore(package): @hoodie/server@21.1.0
BREAKING CHANGE:

Hoodie now keeps state of all databases and access in a dedicated
`hoodie-store` database. If your app uses CouchDB as its backend, there
is no migration required, it happens automatically on next restart.

If you don’t use CouchDB (which is the default), the simplest way to
"migrate" is to delete the `.hoodie/data` folder. It will loose all
data, so this is only an option for local development. If you don’t want
to loose data, you manually have to create the `hoodie-store`
database with documents for each user database. The documents look
like this:

```json
{
   "_id": "db_user/lsdlo55",
   "_rev": "1-a01e6998fb10543ca0402e648dd1d048",
   "access": {
       "read": {
           "role": [
               "id:lsdlo55"
           ]
       },
       "write": {
           "role": [
               "id:lsdlo55"
           ]
       }
   }
}
```